mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9.2",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.4,<0.2.0",
    "langgraph (>=0.4.8,<0.5.0)",
]
